lambdas/requirements.txt
pytz
psycopg2-binary==2.9.7
requests